requirements.txt
Flask==0.10.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
py==1.4.30
pymongo==3.1.1
pytest==2.8.3
python-dateutil==2.4.2
requests==2.8.1
six==1.10.0
Werkzeug==0.11.2
